# docs/swagger.yaml
basePath: /
...
host: localhost:8081
info:
  contact:
    email: frederico.zolio@prefeitura.rio
    name: API Support
    url: https://iplan.rio
  description: API de serviços para aplicativos da Prefeitura do Rio
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: API Go
  version: "1.0"
paths:
...
  /api/public/empregabilidade/vagas:
    get:
      description: Retorna lista paginada de vagas publicadas. Sem filtro retorna
        todos os status públicos (publicado_ativo, publicado_expirado, vaga_congelada,
        vaga_descontinuada). Use ?status= para filtrar por um status específico.
      parameters:
      - description: 'Número da página (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Tamanho da página (default: 10)'
        in: query
        name: pageSize
        type: integer
      - description: Filtrar por status (publicado_ativo, publicado_expirado, vaga_congelada,
          vaga_descontinuada)
        in: query
        name: status
        type: string
      - description: Filtrar por data de publicação (hoje, ultima_semana, ultimo_mes)
        in: query
        name: data_publicacao
        type: string
      - description: 'UUID(s) do regime de contratação (CSV, ex: uuid1,uuid2)'
        in: query
        name: id_regime_contratacao
        type: string
      - description: 'UUID(s) do modelo de trabalho (CSV, ex: uuid1,uuid2)'
        in: query
        name: id_modelo_trabalho
        type: string
      - description: 'Nome(s) ou CNPJ(s) do contratante (CSV, ex: TechRio,12345678000100)'
        in: query
        name: contratante
        type: string
      - description: 'Acessibilidade PCD (CSV, ex: para_pcd,exclusivo_pcd)'
        in: query
        name: acessibilidade_pcd
        type: string
      - description: 'Bairro(s) — busca parcial (CSV, ex: Centro,Tijuca)'
        in: query
        name: bairro
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Listar vagas públicas
      tags:
      - empregabilidade-vagas-public
  /api/public/empregabilidade/vagas/{id}:
    get:
      description: Retorna uma vaga publicada pelo ID (publicado_ativo, publicado_expirado,
        vaga_congelada, vaga_descontinuada)
      parameters:
      - description: ID da vaga
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/empregabilidade.Vaga'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Buscar vaga pública
      tags:
      - empregabilidade-vagas-public
  /api/public/empregabilidade/vagas/slug/{slug}:
    get:
      description: Retorna uma vaga publicada pelo slug (apenas se estiver ativa).
        Slug no formato "{titulo-slugificado}-{12-chars-hex-do-uuid}". Se o slug for
        histórico (título mudou), retorna 301 redirect para o slug atual.
      parameters:
      - description: 'Slug da vaga (ex: analista-financeiro-jr-b06235c80d2a)'
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/empregabilidade.Vaga'
        "301":
          description: Moved Permanently
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Buscar vaga pública por slug
      tags:
      - empregabilidade-vagas-public
...
  /api/v1/empregabilidade/vagas:
    get:
      description: Retorna lista paginada de vagas com filtros opcionais
      parameters:
      - description: 'Número da página (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Tamanho da página (default: 10, máx: 100)'
        in: query
        name: pageSize
        type: integer
      - description: Filtrar por status (em_edicao, em_aprovacao, publicado_ativo,
          publicado_expirado, vaga_congelada, vaga_descontinuada)
        in: query
        name: status
        type: string
      - description: Filtrar por CNPJ do contratante
        in: query
        name: contratante
        type: string
      - description: Filtrar por ID do órgão parceiro (ignorado se middleware já restringiu)
        in: query
        name: orgao_parceiro_id
        type: string
      - description: Busca parcial no título da vaga
        in: query
        name: search
        type: string
      - description: Filtrar por data de publicação (hoje, ultima_semana, ultimo_mes)
        in: query
        name: data_publicacao
        type: string
      - description: 'UUID(s) do regime de contratação (CSV, ex: uuid1,uuid2)'
        in: query
        name: id_regime_contratacao
        type: string
      - description: 'UUID(s) do modelo de trabalho (CSV, ex: uuid1,uuid2)'
        in: query
        name: id_modelo_trabalho
        type: string
      - description: 'Acessibilidade PCD (CSV, ex: para_pcd,exclusivo_pcd)'
        in: query
        name: acessibilidade_pcd
        type: string
      - description: 'Bairro(s) — busca parcial (CSV, ex: Centro,Tijuca)'
        in: query
        name: bairro
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Listar vagas
      tags:
      - empregabilidade-vagas
    post:
      consumes:
      - application/json
      description: Cria uma nova vaga
      parameters:
      - description: Dados da vaga
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/empregabilidade.Vaga'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/empregabilidade.Vaga'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Criar vaga
      tags:
      - empregabilidade-vagas
...
  /api/v1/empregabilidade/vagas/{id}/discontinue:
    put:
      description: Descontinua uma vaga publicada ou congelada e atualiza o status
        de todas as candidaturas para vaga_descontinuada
      parameters:
      - description: ID da vaga
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Descontinuar vaga
      tags:
      - empregabilidade-vagas
...
  /api/v1/empregabilidade/vagas/{id}/freeze:
    put:
      description: Congela uma vaga publicada e atualiza o status de todas as candidaturas
        vinculadas para vaga_congelada
      parameters:
      - description: ID da vaga
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Congelar vaga
      tags:
      - empregabilidade-vagas
...
schemes:
- http
- https
security:
- BearerAuth: []
securityDefinitions:
  BearerAuth:
    description: Autenticação usando token JWT no formato 'Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"